created __init_.py, changed the module name and cleared setup.py
