+++ Tests/test_EBI_dbfetch.py
@@ -7,7 +7,7 @@
 # Builtins
 import unittest
 
-from Bio.EBI import db_info, dbfetch
+from Bio.EBI.debfetch import db_info, dbfetch
 from Bio import SeqIO
